New unfinished custom chunk provider.
This is meant as a replacement for the ChunkProviderServer class, to allow fast loading of chunks. Although the class is nearly finished, I still have to find a way to make the server use it (maybe reflection? although I really don't like that idea. Maybe javaassist might be a better option).